Enables source container image builds
I am only enabling this on push, despite the docs saying it should be enabled on both push and on-pull. It seems very wasteful to have this job running on every PR.

If it starts failing for some reason after merging to main branches, we can consider enabling on the pull request as well.

Signed-off-by: Yorgos Saslis <yorgos.saslis@redhat.com>